requirements.txt
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
numpy>=1.24.3
pandas>=2.1.1
matplotlib>=3.8.0
scikit-learn>=1.3.1
seaborn>=0.13.0
tqdm>=4.66.1
pillow>=10.0.1
